+++ src/helpers/DbHelper.php
@@ -231,14 +231,7 @@ class DbHelper
 
 		if (isset($config['default']) && (is_string($config['default']) || is_bool($config['default']) || is_numeric($config['default'])))
 		{
-			if (is_string($config['default']) && !is_numeric($config['default']))
-			{
-				$def .= ' DEFAULT "'.$config['default'].'"';
-			}
-			else
-			{
-				$def .= ' DEFAULT '.(int)$config['default'];
-			}
+			$def .= ' DEFAULT '.craft()->db->quoteValue($config['default']);
 		}
 
 		if (!empty($config['primaryKey']))